Move classes in it's own files
