+++ src/MahApps.Metro/MahApps.Metro.Shared/Controls/MetroThumbContentControlDragCompletedEventArgs.cs
@@ -1,0 +1,13 @@
+using System.Windows.Controls.Primitives;
+
+namespace MahApps.Metro.Controls
+{
+    public class MetroThumbContentControlDragCompletedEventArgs : DragCompletedEventArgs
+    {
+        public MetroThumbContentControlDragCompletedEventArgs(double horizontalOffset, double verticalOffset, bool canceled)
+            : base(horizontalOffset, verticalOffset, canceled)
+        {
+            this.RoutedEvent = MetroThumbContentControl.DragCompletedEvent;
+        }
+    }
+}
