+++ src/MahApps.Metro/MahApps.Metro.Shared/MahApps.Metro.Shared.projitems
@@ -95,12 +95,16 @@
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroAnimatedTabControl.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroContentControl.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroHeader.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroHeaderAutomationPeer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroNavigationWindow.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroProgressBar.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroTabControl.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroTabItem.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroThumb.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroThumbContentControl.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroThumbContentControlAutomationPeer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroThumbContentControlDragCompletedEventArgs.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroThumbContentControlDragStartedEventArgs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroWindow.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MetroWindowHelpers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\MultiFrameImage.cs" />
